refactor(db-scripts): extract into `unfound_database_script.rb`
